refactor: moved document pages to one folder
