+++ front-doc/src/App.tsx
@@ -5,10 +5,10 @@ import { Route, Routes } from "react-router-dom";
 import { HomePage } from "./pages/home/HomePage";
 import { HeaderComponent } from "./components/header/HeaderComponent";
 import { FooterComponent } from "./components/footer/FooterComponent";
-import { ReportPage } from "./pages/ReportPage";
-import { ReleasePage } from "./pages/ReleasePage";
-import { OrderPage } from "./pages/OrderPage";
-import { ListPage } from "./pages/ListPage";
+import { ReportPage } from "./pages/documentPages/ReportPage";
+import { ReleasePage } from "./pages/documentPages/ReleasePage";
+import { OrderPage } from "./pages/documentPages/OrderPage";
+import { ListPage } from "./pages/documentPages/ListPage";
 
 function App() {
   return (
