+++ front-doc/src/pages/documentPages/ReleasePage.tsx
