+++ front-doc/src/pages/documentPages/ReportPage.tsx
